controller method permission set

+++ app/Http/Controllers/Backend/Blog/CategoryController.php
@@ -6,6 +6,7 @@ use App\Traits\ResponseData;
 use Illuminate\Http\Request;
 use App\Services\CategoryService;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\CategoryRequest;
 
 class CategoryController extends Controller
@@ -20,6 +21,9 @@ class CategoryController extends Controller
     }
 
     public function index(Request $request){
+        //authorized 403
+        Gate::authorize('category-access');
+
         if($request->ajax()){
             return $this->category->allData($request);
         }
@@ -31,25 +35,41 @@ class CategoryController extends Controller
 
     public function storeOrUpdate(CategoryRequest $request){
         if($request->ajax()){
-            return $this->category->storeOrUpdateData($request);
+            if(permission('category-create') || permission('category-edit')){
+                return $this->category->storeOrUpdateData($request);
+            }else{
+                return $this->responseJson('error',UNAUTHORIZED_MSG);
+            }
         }
     }
 
     public function edit(Request $request){
         if($request->ajax()){
-            return $this->category->editData($request->id);
+            if(permission('category-edit')){
+                return $this->category->editData($request->id);
+            }else{
+                return $this->responseJson('error',UNAUTHORIZED_MSG);
+            }
         }
     }
 
     public function delete(Request $request){
         if($request->ajax()){
-            return $this->category->deleteData($request->id);
+            if(permission('category-delete')){
+                return $this->category->deleteData($request->id);
+            }else{
+                return $this->responseJson('error',UNAUTHORIZED_MSG);
+            }
         }
     }
 
     public function statusChange(Request $request){
         if($request->ajax()){
-            return $this->category->statusData($request->id, $request->status);
+            if(permission('category-status')){
+                return $this->category->statusData($request->id, $request->status);
+            }else{
+                return $this->responseJson('error',UNAUTHORIZED_MSG);
+            }
         }
     }
 }
